requirements.txt
alembic==1.5.8
aniso8601==9.0.1
attrs==20.3.0
click==7.1.2
Flask==1.1.2
Flask-JWT-Extended==4.1.0
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
flask-restx==0.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.24.2
psycopg2==2.8.6
PyJWT==2.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.16.0
python-editor==1.0.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.4.4
Werkzeug==1.0.1
